+++ packages/auto-id/__tests__/auto-id.test.tsx
@@ -2,7 +2,7 @@
 /// <reference types="vitest-dom/extend-expect" />
 
 import * as React from "react";
-import { render, cleanup } from "@reach-fork-internal/test/utils";
+import { render, cleanup } from "@reach-fork-hamtolchu-internal/test/utils";
 import { vi, it, expect, describe, afterEach } from "vitest";
 
 const { useId } = await vi.importActual<typeof import("../src/auto-id")>(