+++ packages/combobox/__tests__/combobox.test.tsx
@@ -7,7 +7,7 @@ import {
 	render,
 	withMarkup,
 	userEvent,
-} from "@reach-fork-internal/test/utils";
+} from "@reach-fork-hamtolchu-internal/test/utils";
 import type { ComboboxInputProps } from "@reach/combobox";
 import {
 	Combobox,