+++ packages/descendants/src/reach-descendants.tsx
@@ -3,7 +3,7 @@ import {
 	useForceUpdate,
 	useIsomorphicLayoutEffect as useLayoutEffect,
 	noop,
-} from "@reach-fork/utils";
+} from "@reach-fork-hamtolchu/utils";
 
 declare const __DEV__: boolean;