+++ packages/dialog/__tests__/dialog-axe.test.tsx
@@ -4,7 +4,7 @@
 import * as React from "react";
 import { axe } from "vitest-axe";
 import type { AxeCore } from "vitest-axe";
-import { cleanup, render, act } from "@reach-fork-internal/test/utils";
+import { cleanup, render, act } from "@reach-fork-hamtolchu-internal/test/utils";
 import { Dialog } from "@reach/dialog";
 import { expect, describe, afterEach, it, vi } from "vitest";