+++ packages/disclosure/src/reach-disclosure.tsx
@@ -22,9 +22,9 @@ import {
 	useComposedRefs,
 	composeEventHandlers,
 	useControlledState,
-} from "@reach-fork/utils";
-import type * as Polymorphic from "@reach-fork/polymorphic";
-import { useId } from "@reach-fork/auto-id";
+} from "@reach-fork-hamtolchu/utils";
+import type * as Polymorphic from "@reach-fork-hamtolchu/polymorphic";
+import { useId } from "@reach-fork-hamtolchu/auto-id";
 
 const [DisclosureProvider, useDisclosureCtx] =
 	createContext<DisclosureContextValue>("Disclosure");