+++ packages/dropdown/src/reach-dropdown.tsx
@@ -2,8 +2,8 @@
 /// <reference types="vitest-dom/extend-expect" />
 
 import * as React from "react";
-import { useId } from "@reach-fork/auto-id";
-import { Popover } from "@reach-fork/popover";
+import { useId } from "@reach-fork-hamtolchu/auto-id";
+import { Popover } from "@reach-fork-hamtolchu/popover";
 import {
 	createDescendantContext,
 	DescendantProvider,
@@ -11,7 +11,7 @@ import {
 	useDescendants,
 	useDescendantsInit,
 	useDescendantKeyDown,
-} from "@reach-fork/descendants";
+} from "@reach-fork-hamtolchu/descendants";
 import {
 	composeEventHandlers,
 	createContext,
@@ -23,11 +23,11 @@ import {
 	usePrevious,
 	useStatefulRefValue,
 	makeId,
-} from "@reach-fork/utils";
-import type * as Polymorphic from "@reach-fork/polymorphic";
+} from "@reach-fork-hamtolchu/utils";
+import type * as Polymorphic from "@reach-fork-hamtolchu/polymorphic";
 
-import type { Descendant } from "@reach-fork/descendants";
-import type { Position } from "@reach-fork/popover";
+import type { Descendant } from "@reach-fork-hamtolchu/descendants";
+import type { Position } from "@reach-fork-hamtolchu/popover";
 
 ////////////////////////////////////////////////////////////////////////////////
 // Actions