+++ packages/listbox/__tests__/listbox.test.tsx
@@ -8,7 +8,7 @@ import {
 	act,
 	fireEvent,
 	keyType,
-} from "@reach-fork-internal/test/utils";
+} from "@reach-fork-hamtolchu-internal/test/utils";
 import type { ListboxProps } from "@reach/listbox";
 import {
 	Listbox,