+++ packages/menu-button/__tests__/menu-button-axe.test.tsx
@@ -7,7 +7,7 @@ import {
 	render,
 	screen,
 	simulateMouseClick,
-} from "@reach-fork-internal/test/utils";
+} from "@reach-fork-hamtolchu-internal/test/utils";
 import { axe } from "vitest-axe";
 import {
 	Menu,