+++ packages/visually-hidden/__tests__/visually-hidden.test.tsx
@@ -4,7 +4,7 @@
 import * as React from "react";
 import { VisuallyHidden } from "@reach/visually-hidden";
 import { describe, it, expect, afterEach } from "vitest";
-import { render, cleanup } from "@reach-fork-internal/test/utils";
+import { render, cleanup } from "@reach-fork-hamtolchu-internal/test/utils";
 
 afterEach(cleanup);