+++ packages/window-size/src/reach-window-size.tsx
@@ -9,7 +9,7 @@ import * as React from "react";
 import {
 	canUseDOM,
 	useIsomorphicLayoutEffect as useLayoutEffect,
-} from "@reach-fork/utils";
+} from "@reach-fork-hamtolchu/utils";
 
 ////////////////////////////////////////////////////////////////////////////////